### 2021-08-09 #3595 @whedon
[PRE REVIEW]: spant: An R package for magnetic resonance spectroscopy analysis
**Submitting author:** @martin3141 (<a href="http://orcid.org/0000-0002-2089-3956">Martin Wilson</a>)
**Repository:** <a href="https://github.com/martin3141/spant" target ="_blank">https://github.com/martin3141/spant</a>
**Version:** v1.14.0
