### 2021-08-10 b45aade
.basis file reader unit tests
