### 2021-08-11 0ecd418
lcm basis reading code update
