### 2021-08-23 #3646 @whedon
[REVIEW]: spant: An R package for magnetic resonance spectroscopy analysis
**Submitting author:** @martin3141 (<a href="http://orcid.org/0000-0002-2089-3956">Martin Wilson</a>)
**Repository:** <a href="https://github.com/martin3141/spant" target ="_blank">https://github.com/martin3141/spant</a>
**Version:** v1.15.0

**Note from the Editor:** Due to a reviewer dropping out of the process, I did some aspects of the review - to verify some of the points raised during the process. This is primarily related to reading the paper for comprehensiveness, checking the repository, and installing the software (as issues were raised in that regard).

** Note: ** Functionality and performance were not checked as these were not raised as issues in the peer review.

** Note: ** Other aspects were not raised as potential issues in the review and were not checked.
