### 2021-08-23 #3595 @danielskatz
@a3sha2 - I'm not sure what are you trying to do.  The review is taking place in https://github.com/openjournals/joss-reviews/issues/3646 now.  And this issue for the pre-review should remain closed, since the "action" has moved to the review issue.
