### 2021-09-18 #3646 @joaomcteixeira
First of all, congratulations on this solid and very promising software package.

I have gone through the paper, the repository, and the documentation. I am highlighting here some issues and suggestions I believe can improve the project and should be addressed before accepting the paper.

The article reads clearly. The English level is good, and there are no faults. The repository looks well organized. Also, the documentation pages are easy to navigate and contain useful information, and the API seems well documented.

I found some issues that could be corrected to clarify the project purposes and improve users and developers engagement with the software.

- [ ] It is not clear from the paper nor the documentation if `spant` can process raw FID data or if it works only on preprocessed data, that is, chemical shift vs intensity spectra. This should be clearly stated in the paper and the documentation.
- [ ] the paper states `spant` can process "large and complex datasets". What are large and complex datasets? A short explanation on the nature of these datasets and how they differ from "normal" datasets should be given.
- [ ] It would be beneficial for the paper if it states that `spant` is already available on the MRSHub webpage. This gives value to `spant` because it shows the community has accepted it already.
- [ ] In the documentation page, the tab "Articles" could read "Tutorials" instead.
- [ ] it is not clear the version scheme adopted by the project. Is it SemanticVersion 2.0?
- [ ] there are no community guidelines on how to contribute to the project. For example, how to execute tests locally at the developer level, nor explanation on the tests will run at the repository level upon PR. Or how the CI, if it exists, is managed. Usually, this is stated in the `CONTRIBUTING` file.

I am running Ubuntu 20.04 and following the installation instructions. I do not use R myself, so I had to install R from scratch. R and R-Studio installed well. I then followed the instructions to install `spant` running: `install.packages("spant", dependencies = TRUE)
`. About 30 minutes have already passed, and R is still compiling packages (I have a powerful laptop). Is it supposed to take so long? If so, this should be stated as a warning to avoid the user's frustration. I will complete my review once the installation finishes.

Best,
