### 2021-09-18 martin3141/spant#9 @joaomcteixeira
Installation issues
Hi @martin3141 , I am a reviewer at https://github.com/openjournals/joss-reviews/issues/3646

I am running Ubuntu 20.04 and tried to install R and `spant` following the instructions in the repository.

* installed R with, `sudo apt install r-base`, everything went okay :+1:
* installed R-studio downloading from https://www.rstudio.com/products/rstudio/download/#download (`.deb`), also okay :+1:
* on R-studio I ran `install.packages("spant", dependencies = TRUE)`

it took about 40 minutes to finish the installation and ended with:

```
Warning in install.packages :
  installation of package ‘spant’ had non-zero exit status
```

I closed and reopened R-studio and did and got:

```R
library(spant)
Error in library(spant) : there is no package called ‘spant’
```

Am I doing something wrong?
thanks
