### 2021-09-20 martin3141/spant#9 @martin3141
Hi @joaomcteixeira,

Thanks for taking the time to review this package!

So R compiles packages from source by default on Linux, and yes this can unfortunately time a long time. For Windows and Mac things are much faster because pre-compiled binaries are made available on CRAN.

Your failure was very likely to be a missing system library that one of spant dependencies needs to compile. The missing library will be listed somewhere above the error message, however these messages can be hard to spot.

I just tried installing on a semi-fresh Ubuntu 20.04 install via WSL and needed to install the following:

`sudo apt install libfontconfig1-dev libharfbuzz-dev libfribidi-dev libfreetype6-dev libpng-dev libtiff5-dev libjpeg-dev`

Hopefully the above will solve your problem.

Martin
