+++ README.Rmd
@@ -29,7 +29,7 @@ algorithms are included:
 * Advanced fully-automated metabolite fitting algorithm - ABfit https://onlinelibrary.wiley.com/doi/10.1002/mrm.28385.
 * Robust retrospective frequency and phase correction - RATS https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.27605.
 * Flexible data types to support single voxel, dynamic and spectroscopic imaging data types.
-* Raw data import/export.
+* Raw data import from individual coils and dynamic measurements, eg support for importing individual FIDs from Siemens TWIX formatted data.
 * Publication quality plotting.
 * Extensive set of pre-processing steps (phasing, coil-combination, zero-filling, HSVD filtering...)
 * Quantum mechanical based simulation for experimental design and basis-set generation.
