+++ paper/paper.md
@@ -26,7 +26,7 @@ Magnetic Resonance Spectroscopy (MRS) allows the measurement of small molecules 
 
 Traditional MRS analysis was dominated by the use of proprietary software, either supplied by scanner manufactures or offline tools such as LCModel (@provencher:1993) and jMRUI (@naressi:2001). In more recent years there has been a steadily increasing trend toward the use of open-source methods — with some early examples including TARQUIN (@reynolds:2006; @wilson:2011) and AQSES (@poullet:2007). This trend is set to continue with the recent transition of LCModel to an open-source license, and an acceleration in the development of new open-source methods and packages such as Vespa (@soher:2011), Gannet (@edden:2014), FID-A (@simpson:2017), Osprey (@oeltzschner:2020), suspect (@rowland:2021) and FSL-MRS (@clarke:2021a). The availability of the MRSHub (<https://mrshub.org/>), a new community orientated software sharing and support platform, and the recent development of the NIfTI MRS file format (@clarke:2021b), to aid data sharing and interoperability, are set to further enhance the ecosystem of open-source MRS analysis tools.
 
-The vast majority of recently developed open-source MRS analysis tools have been written in either MATLAB or Python. Whilst all languages have strengths and weaknesses, R is particularly suited to the interactive exploration and batch processing of large and complex datasets — typical of MRS and neuroimaging studies. The `spant` package was developed to combine traditional  and modern MRS data processing techniques with strengths of R, including: plotting/visualization, statistics, machine learning and data wrangling. Furthermore, `spant` may be used to conveniently combine MRS results with other imaging modalities, due to the availability of a wide range of R packages focused on image processing (@muschelli:2019) and support for the NIfTI data format (@whitcher:2011; @clayden:2021).
+The vast majority of recently developed open-source MRS analysis tools have been written in either MATLAB or Python. Whilst all languages have strengths and weaknesses, R is particularly suited to the interactive exploration and batch processing of large and complex datasets — typical of MRS and neuroimaging studies. The `spant` package was developed to combine traditional and modern MRS data processing techniques with strengths of R, including: plotting/visualization, statistics, machine learning and data wrangling. Furthermore, `spant` may be used to conveniently combine MRS results with other imaging modalities, due to the availability of a wide range of R packages focused on image processing (@muschelli:2019) and support for the NIfTI data format (@whitcher:2011; @clayden:2021). `spant` also supports the majority of common MR vendor data formats allowing complete pipelines to be developed, from raw time-domain samples to metabolite quantities derived from spectral fitting.
 
 At the time of writing, `spant` has been used to develop and validate two new MRS spectroscopy analysis algorithms: RATS (@wilson:2019b) and ABfit (@wilson:2021), and has also been used to study cancer (@franco:2021), Alzheimer’s Disease (@montal:2021) and psychosis (@fisher:2020) — confirming its suitability for both MRS methods research and clinical studies.
 
