+++ CONTRIBUTING.md
@@ -1,0 +1,16 @@
+# Contributing to spant development
+
+Contributions to spant are most welcome and generally take the form of bug reports/feature requests, documentation and suggestions for code changes via a pull request.
+
+## Bug reports/feature requests
+
+Please report bugs and make feature requests via [GitHub](https://github.com/martin3141/spant/issues/).
+
+When filing an issue, the most important thing is to include a minimal 
+reproducible example so that we can quickly verify the problem, and then figure out how to fix it.
+
+## Documentation
+
+## Pull requests
+
+http://adv-r.had.co.nz/Style.html
