### 2021-09-20 #3646 @martin3141
Thank you @joaomcteixeira for taking the time to review this work and your encouraging and helpful comments. Please find my responses below:

> It is not clear from the paper nor the documentation if spant can process raw FID data or if it works only on preprocessed data, that is, chemical shift vs intensity spectra. This should be clearly stated in the paper and the documentation.

"Raw data import/export." was a bulleted item in the Overview section. This has been expanded to emphasise the point. The paper has been updated to explain that raw data can be imported.

> the paper states spant can process "large and complex datasets". What are large and complex datasets? A short explanation on the nature of these datasets and how they differ from "normal" datasets should be given.

Text has been added to emphasize the multidimensional nature of modern neuroimaging studies incorporating MRS.

> It would be beneficial for the paper if it states that spant is already available on the MRSHub webpage. This gives value to spant because it shows the community has accepted it already.

The paper has been amended to include this suggestion.

> In the documentation page, the tab "Articles" could read "Tutorials" instead.

I agree this is clearer and have made the suggested change.

> it is not clear the version scheme adopted by the project. Is it SemanticVersion 2.0?

The format is “major.minor.patch.dev”, so yes I believe this is SemanticVersion 2.0.

> there are no community guidelines on how to contribute to the project. For example, how to execute tests locally at the developer level, nor explanation on the tests will run at the repository level upon PR. Or how the CI, if it exists, is managed. Usually, this is stated in the CONTRIBUTING file.

A CONTRIBUTING.md file had been added to the repository as suggested.
