### 2021-09-21 a4029b3
added head and tail functions for dynamic data
