+++ man/get_head_dyns.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/mrs_data_proc.R
+\name{get_head_dyns}
+\alias{get_head_dyns}
+\title{Return the first scans of a dynamic series.}
+\usage{
+get_head_dyns(mrs_data, n = 1)
+}
+\arguments{
+\item{mrs_data}{dynamic MRS data.}
+
+\item{n}{the number of dynamic scans to return.}
+}
+\value{
+first scans of a dynamic series.
+}
+\description{
+Return the first scans of a dynamic series.
+}
