+++ NAMESPACE
@@ -129,6 +129,7 @@ export(get_fh_dyns)
 export(get_fit_map)
 export(get_fp)
 export(get_guassian_pulse)
+export(get_head_dyns)
 export(get_lcm_cmd)
 export(get_metab)
 export(get_mol_names)
@@ -144,6 +145,7 @@ export(get_sh_dyns)
 export(get_slice)
 export(get_subset)
 export(get_svs_voi)
+export(get_tail_dyns)
 export(get_td_amp)
 export(get_tqn_cmd)
 export(get_uncoupled_mol)
