### 2021-09-28 #3646 @joaomcteixeira
Hi @martin3141 , thanks for addressing my points. The only thing remaining from me is to try to install the package. Sorry I am really overwhelmed now. I will try to address this asap. Apart from that. Congratulations on the big work :-)

FYI, be mindful with the rules of SemanticVersioning 2.0, they are very useful when followed: https://semver.org/
