### 2021-10-12 martin3141/spant#9 @joaomcteixeira
Hi,

But was an Ubuntu installation or a "ubuntu on windows" kind-of installation?

I installed R with anaconda, and inside anaconda-R I run `install.packages("spant", dependencies = TRUE)`, and still the same issue, a forever installation. A possible solution would be to list the dependencies in a anaconda requirement file. What are exactly all the dependencies `spant` needs from CRAN? It is rare if those are not in Anaconda. We can try to build that together.
