### 2021-10-12 martin3141/spant#9 @martin3141
Something else to try is : `install.packages("spant")` as this will not install some of the less important dependencies.
