### 2021-10-12 martin3141/spant#9 @joaomcteixeira
Hi @martin3141

I could install `spant` by installing the libraries manually with anaconda. It works very nice as expected when using conda. This is what I did:

1. just in case, I installed what you told me on the global scope: `sudo apt install libfontconfig1-dev libharfbuzz-dev libfribidi-dev libfreetype6-dev libpng-dev libtiff5-dev libjpeg-dev`
2. created a new Anaconda env for R following these instructions [link](https://docs.anaconda.com/anaconda/user-guide/tasks/using-r-language/#creating-an-environment-with-r)
3. I installed the minimum required packaged for `spant`, which are:

```bash
conda activate r_env
conda install -c r r-nloptr
conda install -c conda-forge r-rniftyreg
conda install -c conda-forge r-ptw
```

I execute a couple of your examples in the Jupyter Notebook and they work great.

Then, I tried to create an Anaconda environment considering all `spant` dependencies, here's the result:

```yml
name: spant
channels:
    - conda-forge
    - r
dependencies:
    - r-base
    - r-essentials
    - r-abind
    - r-plyr
    - r-foreach
    - r-pracma
    - r-stringr
    - r-signal
    - r-matrixcalc
    - r-minpack.lm
    - r-nnls
    - r-r.utils
    - r-graphicsqc
    - r-expm
    - r-smoother
    - r-readr
    - r-magrittr
    - r-ptw
    - r-mmand
    - r-RNifti
    - r-RNiftyReg
    - r-fields
    - r-MASS
    - r-numDeriv
    - r-nloptr
    - r-irlba
    - r-tibble
    - r-jsonlite
    - r-viridisLite
    - r-shiny
    - r-miniUI
    - r-knitr
    - r-rmarkdown
    - r-testthat
    - r-ragg
    - r-doParallel
    #- r-complexplus
    #- r-grDevices
```

I could not find on Anaconda repositories the `complexplus` package. The `grDevices` can be replaced by [farver](https://github.com/thomasp85/farver), I think. But I believe that would require you to change the `spant` code in some parts.

Anyway, with the configuration above, even without these two packages, the Anaconda environment builds easily, and after, the command:

```r
install.packages("spant")
```

works very nice also.

These anaconda instructions would work for Win, Mac, and Linux alike (in principle).

The final instructions would be (after installing Anaconda):

```
conda env create -f spant_requirements.yml
conda activate spant
```

Where `spant_requirements.yml` is a file with the contents of the dependencies as I pasted above. You can change the name of the environment; here I named it `spant` for simplicity. It is up to the user to install RStudio (also available in Anaconda).

I think you can add these instructions to your package. They would help you a lot getting new users.

Cheers,
