### 2021-10-12 #3646 @joaomcteixeira
@chartgerink

I have proposed @martin3141 installations instructions for Anaconda. Read here: https://github.com/martin3141/spant/issues/9

I think @martin3141 can add those instructions to the package. However, for me it is not a problem if Martin decides not to add them. This should not hinder publication, however it would considerably enhance the dissemination of the project. Answering the question: "Does the package installs correctly?" @martin3141 said he could install it on Ubuntu, but I couldn't. I can only install it and use it with the Anaconda-based approach. With the Anaconda env it seems to work nicely. I executed a couple of the official examples in a Jupyter Notebook and everything went accordingly.

Unless there is any other pressing issue you want me to address, I conclude here my revision of the `spant` package, and I support its publication.

Congratulations @martin3141 for your nice work :tada:
