### 2021-10-12 d459b5e
updated installed instructions for ubuntu 20.04
