+++ README.Rmd
@@ -52,6 +52,15 @@ install.packages("devtools")
 devtools::install_github("martin3141/spant", ref = "devel")
 ```
 
+### Ubuntu 20.04 installation
+CRAN packages need to be compiled on Linux, and therefore you may need to ensure some additional system libraries are installed. spant may be installed from a clean installation of Ubuntu 20.04 with the following commands pasted into the terminal:
+
+```{ubuntu install, eval = FALSE}
+sudo apt install -y r-base libcurl4-openssl-dev libfontconfig1-dev libharfbuzz-dev libfribidi-dev libtiff5-dev
+mkdir -p ~/R/x86_64-pc-linux-gnu-library/3.6
+Rscript -e 'install.packages("spant", dependencies = TRUE)'
+```
+
 ## Documentation
 Quick introduction to the basic analysis workflow : https://martin3141.github.io/spant/articles/spant-intro.html
 
