### 2021-10-12 martin3141/spant#9 @martin3141
Hi @joaomcteixeira,

Thanks for persisting with the anaconda build. I haven't used R from within this environment before, but I'm sure you are right that others may find the instructions helpful. I'll add these to the README.md once I've been able to confirm the recipe works on my system too.

I made a fresh Ubuntu 20.04 VM to test the install, and have added instructions to the README.md for users on this platform wanting a traditional (not anaconda) install.

Cheers,

Martin
