### 2021-10-13 martin3141/spant#9 @joaomcteixeira
Perfect, up to you now :wink:
I was the reviewer of your article but I am also a developer, so I know this kind of help is very useful. You can always have a file with different installation instructions for all the different flavors you wish to support. It all depends on your community demands. Everyone coming from python will likely use anaconda.

you can close the issue,
Best for your project :+1:
