### 2021-10-16 #3646 @chartgerink
Dear @martin3141 👋

Thanks for bearing with us on this review process of your package. It has taken a bit longer than I had hoped - and I am also getting the hang of this as a first-time editor.

The paper presents an important package for R environments, which up to now has been more readily available in Python environments.

Based on the review, I wanted to ensure the installation worked on Linux platforms; I attempted the installation (https://github.com/martin3141/spant/issues/9) on a new Ubuntu instance with the new instructions. This did not work (FWIW, the error was `ERROR: dependency ‘RNiftyReg’ is not available for package ‘spant’` and trying to install that led to the error `ERROR: compilation failed for package ‘RcppEigen’`).

As the installation is a remaining issue, this means I cannot yet accept the paper. I know you cannot control the user's machine, but there needs to be a clear way to install the software on Linux and its relevant dependencies. Please test on an empty install and verify the instructions are complete.

Below I outline the todo's for your revisions from my end. If you can resolve these, I am happy to accept the paper for publication.

## Todo

- [ ] Check [your references based on Whedon's suggestions](https://github.com/openjournals/joss-reviews/issues/3646#issuecomment-944873756)
- [ ] Please test the installation for Ubuntu - it seems like this is a pernicious issue and will prevent usage of your software
- [ ] Please check and fix the failing issues on R-CMD-CHECK - fixing this will help the reception of your package (as it may gain more traffic upon publication)
