+++ .github/workflows/R-CMD-check.yaml
@@ -25,7 +25,6 @@ jobs:
           - {os: ubuntu-20.04,   r: 'oldrel-1'}
           - {os: ubuntu-18.04,   r: 'release'}
           - {os: ubuntu-18.04,   r: 'oldrel-1'}
-          - {os: ubuntu-21.10,   r: 'release'}
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
