### 2021-10-16 21c102f
added Ubuntu 21.10 install instructions
