### 2021-10-16 2cb1e11
updated paper doi's
