+++ paper/paper.bib
@@ -17,6 +17,7 @@
   volume  = {56},
   number  = {6},
   pages   = {1211--1219},
+  doi     = {10.1002/mrm.21081},
   month   = {Dec}
 }
 
@@ -28,6 +29,7 @@
   volume  = {65},
   number  = {1},
   pages   = {1--12},
+  doi     = {10.1002/mrm.22579},
   month   = {Jan}
 }
 
@@ -39,6 +41,7 @@
   volume  = {34},
   number  = {5},
   pages   = {e4257},
+  doi     = {10.1002/nbm.4257},
   month   = {May}
 }
 
@@ -50,6 +53,7 @@
   volume  = {30},
   number  = {6},
   pages   = {672--679},
+  doi     = {10.1002/mrm.1910300604},
   month   = {Dec}
 }
 
@@ -61,6 +65,7 @@
   volume  = {40},
   number  = {6},
   pages   = {1445--1452},
+  doi     = {10.1002/jmri.24478},
   month   = {Dec}
 }
 
@@ -80,6 +85,7 @@
   volume  = {85},
   number  = {6},
   pages   = {2950--2964},
+  doi     = {10.1002/mrm.28630},
   month   = {06}
 }
 
@@ -90,6 +96,7 @@
   year    = {2020},
   volume  = {343},
   pages   = {108827},
+  doi     = {10.1016/j.jneumeth.2020.108827},
   month   = {09}
 }
 
@@ -101,6 +108,7 @@
   volume  = {31},
   number  = {4},
   pages   = {269--286},
+  doi     = {10.1016/s0010-4825(01)00006-3},
   month   = {Jul}
 }
 
@@ -112,6 +120,7 @@
   volume  = {20},
   number  = {5},
   pages   = {493--504},
+  doi     = {10.1002/nbm.1112},
   month   = {Aug}
 }
 
@@ -123,6 +132,7 @@
   volume  = {82},
   number  = {2},
   pages   = {527--550},
+  doi     = {10.1002/mrm.27742},
   month   = {08}
 }
 
@@ -134,6 +144,7 @@
   volume  = {270},
   number  = {3},
   pages   = {658--679},
+  doi     = {10.1148/radiol.13130531},
   month   = {Mar}
 }
 
@@ -165,6 +176,7 @@
   volume  = {85},
   number  = {1},
   pages   = {13--29},
+  doi     = {10.1002/mrm.28385},
   month   = {01}
 }
 
@@ -176,6 +188,7 @@
   volume  = {81},
   number  = {5},
   pages   = {2878--2886},
+  doi     = {10.1002/mrm.27605},
   month   = {05}
 }
 
@@ -186,6 +199,7 @@
   year    = {2021},
   volume  = {13},
   number  = {10},
+  doi     = {10.3390/cancers13102417},
   month   = {May}
 }
 
@@ -197,6 +211,7 @@
   volume  = {10},
   number  = {1},
   pages   = {254},
+  doi     = {10.1038/s41398-020-00927-x},
   month   = {07}
 }
 
@@ -224,6 +239,7 @@
   volume  = {44},
   number  = {6},
   pages   = {1--28},
+  doi     = {10.18637/jss.v044.i06},
   url     = {https://www.jstatsoft.org/v44/i06/}
 }
 
