### 2021-10-16 #3646 @martin3141
Dear @chartgerink,

Thanks for your final checks and suggestions.

1) References fixed.
2) I wasn't able to reproduce your problem, but the install commands are slightly different for Ubuntu 21.10, so perhaps this is the cause. I've added instructions for that specific version to the main github page.
3) Checks fixed. Hopefully you can see from : [https://github.com/martin3141/spant/actions/runs/1349259672]() that the software installs and passes internal and standard CRAN checks on the most recent LTS versions of Ubuntu (18.04 and 20.04).

@joaomcteixeira thanks for the tips on Anaconda, I will definitely try this 👍

Martin
