### 2021-10-16 7b399ba
added anaconda install instructions
