### 2021-10-16 cf80047
dropped matrixcalc dependency
