+++ DESCRIPTION
@@ -28,7 +28,6 @@ Imports:
     stringr,
     complexplus,
     signal,
-    matrixcalc,
     minpack.lm,
     nnls,
     utils,
