+++ NAMESPACE
@@ -76,7 +76,6 @@ export(check_tqn)
 export(circ_mask)
 export(collapse_to_dyns)
 export(comb_coils)
-export(comb_csv_results)
 export(comb_fit_list_fit_tables)
 export(comb_fit_list_result_tables)
 export(comb_fit_tables)
