### 2021-10-18 #3646 @martin3141
>
>
> @martin3141
>
> Try to find a way to make installation instructions using Anaconda, as we discussed in the issue. Only one or two the dependencies were not installable. Almost everybody uses Anaconda right now. Those who don't, won't care if you ask them to install Anaconda, and those who do, will actually appreciate you have instructions for Anaconda. You can also try a folder-contained Miniconda installation, such that the Anaconda doesn't propagate to the whole system. Good luck!

Thanks @joaomcteixeira, your instructions worked for me and have been added to the README.md.
