+++ R/varpro_basic.R
@@ -1,0 +1,209 @@
+varpro_basic <- function(y, acq_paras, basis, opts = NULL) {
+  mrs_data <- vec2mrs_data(y, fs = acq_paras$fs, ft = acq_paras$ft, 
+                           ref = acq_paras$ref)
+  
+  # has this data been masked?
+  if (is.na(y[1])) return(list(amps = NA, crlbs = NA, diags = NA, fit = NA))
+  
+  # use default fitting opts if not specified 
+  if (is.null(opts)) opts <- varpro_basic_opts()
+  
+  y <- drop(mrs_data$data)
+  
+  
+  if (opts$method == "td") {
+    proj <- td_projection(y, basis, opts$nnls)
+  } else if (opts$method == "fd") {
+    proj <- fd_projection(y, basis, opts$nnls)
+  } else if (opts$method == "fd_re") {
+    proj <- fd_re_projection(y, basis, opts$nnls)
+  } else {
+    stop("unrecognised varpro method")
+  }
+
+  # create some common metabolite combinations
+  amps <- append_metab_combs(proj$amps)
+  
+  fit <- data.frame(PPMScale = ppm(mrs_data, N = 2 * length(y)),
+                    Data = Re(proj$Y), Fit = Re(proj$YHAT),
+                    Baseline = Re(proj$BL))
+  
+  fit <- cbind(fit, proj$basis_frame)
+  
+  class(fit) <- c("fit_table", "data.frame")
+  
+  diags <- data.frame(dummy = 1, stringsAsFactors = TRUE)
+  
+  list(amps = amps, crlbs = t(rep(NA, length(amps))), diags = diags, fit = fit)
+}
+
+fd_projection <- function(y, basis, nnls) {
+  
+  Npts   <- length(y)
+  Nbasis <- dim(basis$data)[2]
+  
+  Y <- ft_shift(y)
+  Y_real <- c(Re(Y), Im(Y))
+  
+  basis_fd   <- basis$data
+  basis_real <- rbind(Re(basis_fd), Im(basis_fd))
+  
+  if (nnls) {
+    ahat <- nnls(basis_real, Y_real)$x
+  } else {
+    ahat <- stats::.lm.fit(basis_real, Y_real)$coefficients
+  }
+  
+  YHAT <- basis_fd %*% ahat
+  yhat <- ift_shift(as.vector(YHAT))
+  amat <- matrix(ahat, nrow = Npts, ncol = Nbasis, byrow = TRUE)
+  basis_sc <- basis_fd * amat
+  basis_sc <- apply(basis_sc, 2, ift_shift)
+  zero_mat <- matrix(0, nrow = Npts, ncol = Nbasis)
+  basis_sc <- rbind(basis_sc, zero_mat)
+  BASIS_SC <- apply(basis_sc, 2, ft_shift)
+  
+  basis_frame <- as.data.frame(Re(BASIS_SC), row.names = NA)
+  colnames(basis_frame) <- basis$names
+  
+  # zero pad
+  yhat <- c(yhat, rep(0, Npts))
+  YHAT <- ft_shift(as.vector(yhat))
+  
+  # zero pad
+  y <- c(y, rep(0, Npts))
+  Y <- ft_shift(y)
+  BL <- rep(0, length(Y))
+  
+  amps <- data.frame(t(ahat))
+  colnames(amps) <- basis$names
+  
+  return(list(amps = amps, Y = Y, YHAT = YHAT, BL = BL,
+              basis_frame = basis_frame))
+}
+
+fd_re_projection <- function(y, basis, nnls) {
+  
[…]
