+++ man/varpro_basic_opts.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/varpro_basic.R
+\name{varpro_basic_opts}
+\alias{varpro_basic_opts}
+\title{Return a list of options for a basic VARPRO analysis.}
+\usage{
+varpro_basic_opts(method = "fd_re", nnls = TRUE)
+}
+\arguments{
+\item{method}{one of "td", "fd", "fd_re".}
+
+\item{nnls}{restrict basis amplitudes to non-negative values.}
+}
+\value{
+full list of options.
+}
+\description{
+Return a list of options for a basic VARPRO analysis.
+}
