### 2021-11-01 #3646 @kthyng
@martin3141 Please update the Zenodo archive metadata so the title and author list exactly match your JOSS submission.
