+++ paper/paper.bib
@@ -224,7 +224,7 @@
 }
 
 @manual{clayden:2021,
-  title  = {rnifti: fast r and c++ access to nifti images},
+  title  = {rnifti: fast {R} and {C}++ access to nifti images},
   author = {Jon Clayden and Bob Cox and Mark Jenkinson},
   year   = {2021},
   note   = {R package version 1.3.1},
