+++ paper/paper.bib
@@ -194,7 +194,7 @@
 
 @article{franco:2021,
   author  = {Franco, P.  and Huebschle, I.  and Simon-Gabriel, C. P.  and Dacca, K.  and Schnell, O.  and Beck, J.  and Mast, H.  and Urbach, H.  and Wuertemberger, U.  and Prinz, M.  and Hosp, J. A.  and Delev, D.  and Mader, I.  and Heiland, D. H. },
-  title   = {{{M}apping of {M}etabolic {H}eterogeneity of {G}lioma {U}sing {M}{R}-{S}pectroscopy}},
+  title   = {{{M}apping of metabolic heterogeneity of glioma using {M}{R}-spectroscopy}},
   journal = {Cancers (Basel)},
   year    = {2021},
   volume  = {13},
@@ -217,7 +217,7 @@
 
 @article{montal:2021,
   author  = {Montal, V.  and Barroeta, I.  and Bejanin, A.  and Pegueroles, J.  and Carmona-Iragui, M.  and Altuna, M.  and Benejam, B.  and Videla, L.  and Fernández, S.  and Padilla, C.  and Aranha, M. R.  and Iulita, M. F.  and Vidal-Piñeiro, D.  and Alcolea, D.  and Blesa, R.  and Lleó, A.  and Fortea, J. },
-  title   = {{{M}etabolite {S}ignature of {A}lzheimer's {D}isease in {A}dults with {D}own {S}yndrome}},
+  title   = {{{M}etabolite  signature of  alzheimer's disease in adults with down syndrome}},
   journal = {Ann Neurol},
   year    = {2021},
   month   = {Jul}
@@ -244,7 +244,7 @@
 }
 
 @article{clarke:2021b,
-  title  = {NIfTI-MRS format specification},
+  title  = {{NIfTI-MRS} format specification},
   doi    = {10.5281/zenodo.5168794},
   author = {Clarke, W. T. and Wilson, M.},
   year   = {2021},
