### 2021-11-02 #3646 @kthyng
@martin3141 Ok I see the Zenodo update and you capitalized R and C++ in the one reference. However, please look through the rest of the references to update more. I googled "nifti" and apparently it has particular capitalization. "Alzheimer's" is also a proper noun.
