+++ paper/paper.bib
@@ -217,14 +217,14 @@
 
 @article{montal:2021,
   author  = {Montal, V.  and Barroeta, I.  and Bejanin, A.  and Pegueroles, J.  and Carmona-Iragui, M.  and Altuna, M.  and Benejam, B.  and Videla, L.  and Fernández, S.  and Padilla, C.  and Aranha, M. R.  and Iulita, M. F.  and Vidal-Piñeiro, D.  and Alcolea, D.  and Blesa, R.  and Lleó, A.  and Fortea, J. },
-  title   = {{{M}etabolite  signature of  alzheimer's disease in adults with down syndrome}},
+  title   = {{{M}etabolite  signature of  Alzheimer's disease in adults with down syndrome}},
   journal = {Ann Neurol},
   year    = {2021},
   month   = {Jul}
 }
 
 @manual{clayden:2021,
-  title  = {rnifti: fast {R} and {C}++ access to nifti images},
+  title  = {{RNifti}: fast {R} and {C}++ access to {NIfTI} images},
   author = {Jon Clayden and Bob Cox and Mark Jenkinson},
   year   = {2021},
   note   = {R package version 1.3.1},
