+++ paper/paper.bib
@@ -217,7 +217,7 @@
 
 @article{montal:2021,
   author  = {Montal, V.  and Barroeta, I.  and Bejanin, A.  and Pegueroles, J.  and Carmona-Iragui, M.  and Altuna, M.  and Benejam, B.  and Videla, L.  and Fernández, S.  and Padilla, C.  and Aranha, M. R.  and Iulita, M. F.  and Vidal-Piñeiro, D.  and Alcolea, D.  and Blesa, R.  and Lleó, A.  and Fortea, J. },
-  title   = {{{M}etabolite  signature of  Alzheimer's disease in adults with down syndrome}},
+  title   = {{{M}etabolite  signature of  {A}lzheimer's disease in adults with {D}own syndrome}},
   journal = {Ann Neurol},
   year    = {2021},
   month   = {Jul}
